+++ src/main/ruby/truffleruby/core/truffle/kernel_operations.rb
@@ -190,5 +190,19 @@ module Truffle
       end
       [omit, length]
     end
+
+    KERNEL_CLASS = Kernel.instance_method(:class)
+    KERNEL_FROZEN = Kernel.instance_method(:frozen?)
+    private_constant :KERNEL_CLASS, :KERNEL_FROZEN
+
+    # Returns whether the value is frozen, even if the value's class does not include `Kernel`.
+    def self.value_frozen?(value)
+      KERNEL_FROZEN.bind(value).call
+    end
+
+    # Returns the class for the value, even if the value's class does not include `Kernel`.
+    def self.value_class(value)
+      KERNEL_CLASS.bind(value).call
+    end
   end
 end