+++ CHANGELOG.md
@@ -130,6 +130,7 @@ Compatibility:
 * Do not define unsupported C-API TracePoint events to let C extensions detect it (#1983).
 * `IO.copy_stream(in, STDOUT)` now writes to `STDOUT` without buffering like MRI.
 * Implemented `RbConfig['vendordir']`.
+* Implemented `ObjectSpace::WeakMap` (#1385, #1958).
 
 Changes: